+++ DepositCalculator/Models/Currency.cs
@@ -4,11 +4,11 @@ public class Currency : ICurrency
 {
 	public string Name { get; set; }
 
-	public decimal AnnualInterestRate { get; set; }
+	public double AnnualInterestRate { get; set; }
 
 	/// <param name="name">Currency name</param>
 	/// <param name="annualInterestRate">Annual Interest Rate value in %</param>
-	public Currency(string name, decimal annualInterestRate)
+	public Currency(string name, double annualInterestRate)
 	{
 		Name = name;
 		AnnualInterestRate = annualInterestRate;
